Keep only current mode and location as state
Rather than pass transition data around so much, only queue tasks to
update the current mode. When the events trigger, they'll determine the
right mode for that time, and update everything.

This avoids events screwing up in case the location has changed. It also
helps avoids timing issues when the system is suspended, though that'll
actually be improved upon separately.