Set next transition alarm based on what's the next expected transition
Setting the next transition alarm based what's the closest sunrise or
sunset caused this alarm to erroneously be canceled if the current time was
still close to the last transition time.

It also uses a common current time for both determining what should be the
current mode and setting the next transition alarm to avoid inconsistency
between these two.

Fixes the issue #13 for me.