Use thresholds to configure geoclue
This prevents us from having to start/stop geoclue; we just delegate
responsibility for location updates to it and listen for updates.

This simplifies architecture because we don't need to handle polling
location any more, so the location-retrieval can be better modularised;
it doesn't need to receive any messages from the rest of the service, it
only handles incoming messages.