Expose a custom setting to check darkman's status
When the xdg-desktop-portal is using darkman as as settings backend,
it will query darkman for this value and return it.

If the xdp is not using darkman, it will return nothing for the query,
making it clear that it is not using darkman.

This is intended to be used for troubleshooting.